+++ src/main/java/io/kontur/eventapi/job/EnrichmentJob.java
@@ -1,35 +1,30 @@
 package io.kontur.eventapi.job;
 
-import io.kontur.eventapi.enrichment.InsightsApiRequest;
-import io.kontur.eventapi.enrichment.InsightsApiResponse;
-import io.kontur.eventapi.client.KonturAppsClient;
+import io.kontur.eventapi.enrichment.EventEnrichmentTask;
 import io.kontur.eventapi.dao.FeedDao;
 import io.kontur.eventapi.entity.Feed;
 import io.kontur.eventapi.entity.FeedData;
-import io.kontur.eventapi.entity.FeedEpisode;
 import io.micrometer.core.instrument.MeterRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
-import org.wololo.geojson.FeatureCollection;
 
 import java.util.*;
+import java.util.concurrent.CompletableFuture;
 
 import static io.kontur.eventapi.enrichment.InsightsApiRequestBuilder.buildParams;
-import static io.kontur.eventapi.enrichment.InsightsApiRequestBuilder.buildRequest;
-import static io.kontur.eventapi.enrichment.InsightsApiResponseHandler.processResponse;
 
 @Component
 public class EnrichmentJob extends AbstractJob {
 
     private final Logger LOG = LoggerFactory.getLogger(EnrichmentJob.class);
     protected final FeedDao feedDao;
-    private final KonturAppsClient konturAppsClient;
+    private final EventEnrichmentTask eventEnrichmentTask;
 
-    public EnrichmentJob(MeterRegistry meterRegistry, FeedDao feedDao, KonturAppsClient konturAppsClient) {
+    public EnrichmentJob(MeterRegistry meterRegistry, FeedDao feedDao, EventEnrichmentTask eventEnrichmentTask) {
         super(meterRegistry);
         this.feedDao = feedDao;
-        this.konturAppsClient = konturAppsClient;
+        this.eventEnrichmentTask = eventEnrichmentTask;
     }
 
     @Override
@@ -45,46 +40,15 @@ public class EnrichmentJob extends AbstractJob {
     }
 
     protected void enrichFeed(Feed feed) {
+        List<FeedData> events = feedDao.getNotEnrichedEventsForFeed(feed.getFeedId());
+        LOG.info(String.format("%s feed. %s events to enrich", feed.getAlias(), events.size()));
+
         List<String> feedEnrichment = feed.getEnrichment();
         String feedParamsString = buildParams(feedEnrichment);
 
-        List<FeedData> events = feedDao.getNotEnrichedEventsForFeed(feed.getFeedId());
-        LOG.info(String.format("%s feed. %s events to enrich", feed.getAlias(), events.size()));
-        events.forEach(event -> enrichEvent(event, feedEnrichment, feedParamsString));
-    }
-
-    private void enrichEvent(FeedData event, List<String> feedEnrichment, String feedParamsString) {
-        if (needsEnrichment(event.getEventDetails(), event.getGeometries())) {
-            event.setEventDetails(getAnalytics(event.getGeometries(), feedEnrichment, feedParamsString));
-        }
-        for (FeedEpisode episode : event.getEpisodes()) {
-            if (needsEnrichment(episode.getEpisodeDetails(), episode.getGeometries())) {
-                episode.setEpisodeDetails(getAnalytics(episode.getGeometries(), feedEnrichment, feedParamsString));
-            }
-        }
-        event.setEnriched(enriched(event));
-        feedDao.addAnalytics(event);
-    }
-
-    private Map<String, Object> getAnalytics(FeatureCollection geometry, List<String> feedEnrichment, String feedParamsString) {
-        InsightsApiRequest request = buildRequest(geometry, feedParamsString);
-        try {
-            InsightsApiResponse response = konturAppsClient.graphql(request);
-            return processResponse(response, feedEnrichment);
-        } catch (Exception e) {
-            LOG.error(e.getMessage() + "\n" + request.getQuery());
-            return null;
-        }
-    }
-
-    private boolean needsEnrichment(Map<String, Object> details, FeatureCollection geometries) {
-        return (details == null || details.isEmpty()) &&
-                geometries != null && geometries.getFeatures() != null && geometries.getFeatures().length > 0;
-    }
-
-    private boolean enriched(FeedData event) {
-        return !needsEnrichment(event.getEventDetails(), event.getGeometries()) &&
-                event.getEpisodes().stream().noneMatch(episode ->
-                        needsEnrichment(episode.getEpisodeDetails(), episode.getGeometries()));
+        var eventEnrichmentTasks = events.stream()
+                .map(event -> eventEnrichmentTask.enrichEvent(event, feedEnrichment, feedParamsString))
+                .toArray(CompletableFuture[]::new);
+        CompletableFuture.allOf(eventEnrichmentTasks).join();
     }
 }